Demo is working
This check-in is huge but it shows the working two-player split screen effect I've been imagining for a few years now.

The main content of this demo is in the shader, Game.shader, and in player one's viewport code, Viewport.gd.  The rest of the project is mostly boiler plate to get a demo looking decent.